Moved notification methods to notification repository
