+++ src/Infrastructure/CheepRepository.cs
@@ -2,65 +2,18 @@ using Core;
 
 namespace Infrastructure;
 
-using System.Text.RegularExpressions;
 using Microsoft.EntityFrameworkCore;
-using System.Xml.Linq;
-using System.Runtime.CompilerServices;
-using System.Diagnostics.CodeAnalysis;
-using SQLitePCL;
-using Microsoft.EntityFrameworkCore.Query.SqlExpressions;
-using Infrastructure.Migrations;
 
-public class CheepRepository(CheepDbContext context) : ICheepRepository
+public class CheepRepository(CheepDbContext context, INotificationRepository notificationRepository) : ICheepRepository
 {
     public async Task CreateCheep(Cheep newCheep)
     {
-        var qwe = await context.AddAsync(newCheep);
-         
+        var tracker = await context.AddAsync(newCheep);
 
-        // notification logic
-        Author sender = newCheep.Author!;
-
-        // for followers
-        var followers = from author in context.Users 
-                        .Include(a => a.FollowingList)
-                        where author.FollowingList != null && author.FollowingList.Contains(sender)
-                        select author; 
-
-        // need to save changes for us to find out what the ID of the cheep will be 
+        // need to save changes to ensure that the tracker knows the ID of the cheep
         await context.SaveChangesAsync();
-
-        // for tags. NOTE: Tags are prioritized over following notifications 
-        var taggedAuthors = new List<Author>(); 
-        var tagFinderRegex = new Regex("@(\\w+)"); 
-        var matches = tagFinderRegex.Matches(newCheep.Text); 
-        foreach (Match match in matches) {
-            if (match.Groups.Count < 2) { continue; }
-            var authorToTag = await context.Users.FirstOrDefaultAsync(a => a.UserName == match.Groups[1].Value); 
-            if (authorToTag == null) { continue; } // don't require input validation in higher levels 
-            taggedAuthors.Add(authorToTag); // to make sure we don't double notify
-            Notification notif = new Notification{ 
-                cheepID=qwe.Entity.Id,
-                authorID=authorToTag!.Id,
-                tagNotification=true,
-                isNew=true
-            };
-            await context.AddAsync(notif);
-        }
-
-        foreach(var f in followers) {
-            if(taggedAuthors.Contains(f)) { continue; }
-            Notification notif = new Notification{ 
-                cheepID=qwe.Entity.Id,
-                authorID=f.Id,
-                tagNotification=false,
-                isNew=true
-            };
-            await context.AddAsync(notif);
-        }
-
-
-        await context.SaveChangesAsync();
+        
+        await notificationRepository.CreateNotification(tracker);
     }
 
     public async Task<ICollection<Cheep>> ReadCheeps(int limit = -1, int offset = 0, string? authorNameRegex = null)
